+++ packages/theme/utils/constants.ts
@@ -1,0 +1,18 @@
+export const alignments = ["horizontal", "vertical"] as const;
+export const sizes = ["xsm", "sm", "md", "lg", "xlg"] as const;
+export const radiuses = ["none", "sm", "md", "lg", "full"] as const;
+export const variants = [
+  "default",
+  "primary",
+  "secondary",
+  "success",
+  "warning",
+  "error",
+] as const;
+export const appearances = [
+  "solid",
+  "outline",
+  "outline-dashed",
+  "no-outline",
+  "glowing",
+] as const;