+++ packages/theme/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./animations";
 export * from "./colors";
 export * from "./fonts";
 export * from "./shadows";
+export * from "./constants";