+++ packages/button/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aristobyte-ui/button
 
+## 1.0.65
+
+### Patch Changes
+
+- feat: add theme contants
+
+- Updated dependencies []:
+  - @aristobyte-ui/spinner@1.0.65
+  - @aristobyte-ui/utils@1.0.65
+
 ## 1.0.64
 
 ### Patch Changes